# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Shift Scheduler Service API",
        "contact": {
            "name": "Yunus Emre Alpu",
            "url": "https://yunusemrealpu.netlify.app",
            "email": "YunusAlpu@icloud.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/shift-scheduler-service",
    "paths": {
        "/shift-schedule/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update a shift schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift"
                ],
                "summary": "update a shift schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shift Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update shift schedule",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateShiftScheduleDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully updated shift schedule",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "400": {
                        "description": "cannot update shift schedule due to invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "422": {
                        "description": "cannot update shift schedule due to invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "500": {
                        "description": "cannot update shift schedule due to internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.RespondJson": {
            "type": "object",
            "properties": {
                "intent": {
                    "type": "string"
                },
                "message": {},
                "status": {
                    "type": "boolean"
                }
            }
        },
        "handlers.updateShiftScheduleDTO": {
            "type": "object",
            "required": [
                "alias",
                "description",
                "end_date",
                "frequency",
                "manager",
                "organization",
                "shifts",
                "start_date",
                "users",
                "year"
            ],
            "properties": {
                "alias": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "frequency": {
                    "description": "1, 2, 3, 4, 5, 6, 7 (days of the week)",
                    "type": "integer"
                },
                "manager": {
                    "type": "array",
                    "items": {}
                },
                "organization": {
                    "type": "array",
                    "items": {}
                },
                "shifts": {
                    "type": "array",
                    "items": {}
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "description": "0: pending, 1: approved, 2: rejected",
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {}
                },
                "year": {
                    "type": "integer"
                }
            }
        }
    }
}